+++ src/main/kotlin/br/com/devsrsouza/intellij/dropboxfocus/ui/dialog/StartupFocusProjectDialog.kt
@@ -41,6 +41,8 @@ class StartupFocusProjectDialog(
                         focusService.focusOn(settings, module.gradleModulePath)
                         close(0)
                     },
+                    clearFocus = {},
+                    withClearFocusButton = false,
                 )
             }
         }